% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mgf.R
\name{mgfKeyValues}
\alias{mgfKeyValues}
\title{Mgf key/value pairs}
\usage{
mgfKeyValues()
}
\value{
A \code{data.frame} with mgf key/value pairs.
}
\description{
It is possible to
}
\examples{
mgfKeyValues()
}
\author{
Laurent Gatto
}
