% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMgf.R
\name{MsBackendMgf}
\alias{MsBackendMgf}
\alias{MsBackendMgf-class}
\alias{backendInitialize,MsBackendMgf-method}
\title{MS data backend for mgf files}
\usage{
\S4method{backendInitialize}{MsBackendMgf}(object, files, nonStop = FALSE, ..., BPPARAM = bpparam())

MsBackendMgf()
}
\arguments{
\item{object}{Instance of \code{MsBackendMgf} class.}

\item{files}{\code{character} with the (full) file name(s) of the mgf file(s)
from which MS/MS data should be imported.}

\item{nonStop}{\code{logical(1)} whether import should be stopped if an
xml file does not contain all required fields. Defaults to
\code{nonStop = FALSE}.}

\item{...}{Currently ignored.}

\item{BPPARAM}{Parameter object defining the parallel processing
setup to import data in parallel. Defaults to \code{BPPARAM = bpparam()}. See \code{\link[=bpparam]{bpparam()}} for more information.}
}
\description{
The \code{MsBackendMgf} class supports import of MS/MS spectra data from
files in Mascot Generic Format
(\href{http://www.matrixscience.com/help/data_file_help.html}{mgf})
files. After initial import, the full MS data is kept in
memory. \code{MsBackendMgf} extends the \code{\link[=MsBackendDataFrame]{MsBackendDataFrame()}} backend
directly and supports thus the \code{\link[=applyProcessing]{applyProcessing()}} function to make
data manipulations persistent. The backend does however not
support export to mgf files yet.

New objects are created with the \code{MsBackendMgf} function. The
\code{backendInitialize} method has to be subsequently called to
initialize the object and import MS/MS data from (one or more) mgf
files.  Optional parameter \code{nonStop} allows to specify whether the
import returns with an error if one of the xml files lacks required
data, such as \code{mz} and \code{intensity} values (default \code{nonStop = FALSE}), or whether only affected file(s) is(are) skipped and a
warning is shown (\code{nonStop = TRUE}). Note that any other error
(such as xml import error) will abort import regardless of
parameter \code{nonStop}.
}
\examples{

## Create an MsBackendHmdbXml backend and import data from test xml files.
fls <- dir(system.file("extdata", package = "MsBackendMgf"),
    full.names = TRUE, pattern = "mgf$")
be <- backendInitialize(MsBackendMgf(), fls)
be

be$msLevel
be$intensity
be$mz
}
\author{
Laurent Gatto and Johannes Rainer
}
